Move the order components to the components folder
